pyproject.toml
[project]
name = "credo"
version = "0.1.0"
description = ""
authors = [
    {name = "Anonymous",email = "."}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.14.1",
    "scikit-learn>=1.5.1",
    "matplotlib>=3.9.2",
    "seaborn>=0.12.2",
    "tqdm>=4.66.5",
    "torch>=2.3.1",
    "gpytorch>=1.13",
    "pymc",
    "pymc-bart",
    "arviz",
    "pyqt5 (>=5.15.11,<6.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "xlrd (>=2.0.2,<3.0.0)",
    "gpjax (>=0.13.6,<0.14.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "credo"
version = "0.1.0"
description = "Epistemic-Aware Conformalized Credal Envelopes for Regression."
authors = ["Anonymous <.>"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "credo", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
numpy = ">=1.26"
scipy = ">=1.14.1"
scikit-learn = ">=1.5.1"
matplotlib = ">=3.9.2"
seaborn = ">=0.12.2"
tqdm = ">=4.66.5"
torch = ">=2.3.1"
gpytorch = ">=1.13"
pymc = "==5.26.1"
pymc-bart = "==0.11.0"
arviz = "==0.22.0"
pyqt5  = ">=5.15.11,<6.0.0"
requests = ">=2.32.5,<3.0.0"
xlrd = ">=2.0.2,<3.0.0"
gpjax = ">=0.13.6,<0.14.0"

[tool.poetry.group.dev.dependencies]
scikit-learn = "^1.5.1"
matplotlib = "^3.9.2"
catboost = "^1.2.8"
lightgbm = "^4.6.0"
sklearn_quantile = "^0.1.1"
